+++ src/pumpkin/particles.h
@@ -46,25 +46,25 @@ namespace pmk
 
 		PhysicsMaterial* NewPhysicsMaterial();
 
-		void DeletePhysicsMaterial(uint32_t physics_mat_index);
+		void DeletePhysicsMaterial(uint8_t physics_mat_index);
 
 		// Set the physics material's index into render materials. Determines how each physics material is rendered.
-		void SetPhysicsMaterialRender(uint32_t physics_mat_index, uint32_t render_mat_index);
+		void SetPhysicsMaterialRender(uint8_t physics_mat_index, uint32_t render_mat_index);
 
 		// Get the physics material's index into render materials.
-		uint32_t GetPhysicsMaterialRender(uint32_t physics_mat_index);
+		uint32_t GetPhysicsMaterialRender(uint8_t physics_mat_index);
 
 		template<typename T>
-		void SetPhysicsMaterialModel(uint32_t physics_mat_index)
+		void SetPhysicsMaterialModel(uint8_t physics_mat_index)
 		{
 			mpm_context_.SetPhysicsMaterialModel<T>(physics_mat_index);
 		}
 
-		ConstitutiveModel* GetPhysicsMaterialModel(uint32_t physics_mat_index);
+		ConstitutiveModel* GetPhysicsMaterialModel(uint8_t physics_mat_index);
 
-		std::vector<std::pair<float*, std::string>> GetPhysicsParameters(uint32_t physics_mat_index);
+		std::vector<std::pair<float*, std::string>> GetPhysicsParameters(uint8_t physics_mat_index);
 
-		void PhysicsParametersMutated(uint32_t physics_mat_index);
+		void PhysicsParametersMutated(uint8_t physics_mat_index);
 
 #ifdef EDITOR_ENABLED
 		void SetMPMDebugParticleGenEnabled(bool enabled);
